requirement.txt
opencv-python==4.10.0.84
pytesseract==0.3.13
numpy==1.26.4
ultralytics==8.3.5
mysql-connector-python==9.0.0
python-dotenv==1.0.1